Made steps for Repairer registration
Made repairer registration in 3 steps, add a field for itinerant repairers
